LR7/counter-deploy/backend/requirements.txt
Flask==2.2.5
gunicorn==20.1.0
redis==4.6.0
Flask-Cors==3.0.10
python-dotenv==1.0.0
requests==2.31.0
